requirements.txt
pandas
numpy
scikit-learn
joblib
streamlit
extra-streamlit-components
plotly
fpdf
pyjwt
tldextract
requests
fastapi
uvicorn
python-dotenv
